Return EXTRA_SIGN_KEY_ID = 0L for "none" key chosen for ACTION_GET_SIGN_KEY_ID

+++ OpenKeychain/src/main/java/org/sufficientlysecure/keychain/Constants.java
@@ -151,9 +151,9 @@ public final class Constants {
     }
 
     public static final class key {
-        public static final int none = 0;
-        public static final int symmetric = -1;
-        public static final int backup_code = -2;
+        public static final long none = 0;
+        public static final long symmetric = -1;
+        public static final long backup_code = -2;
     }
 
 }
